feat: 결제 로직 변경
